Minor edits.

+++ src/alge/cs_benchmark_matrix.cpp
@@ -731,8 +731,8 @@ _matrix_check(int                          n_variants,
   CS_MALLOC_HD(y, n_cols_ext*6, cs_real_t, cs_alloc_mode);
   BFT_MALLOC(yr0, n_cols_ext*6, cs_real_t);
 
-  CS_MALLOC_HD(da, n_cols_ext*6*6, cs_real_t, cs_alloc_mode);
-  CS_MALLOC_HD(xa, n_edges*2*6*6, cs_real_t, cs_alloc_mode);
+  CS_MALLOC_HD(da, (size_t)n_cols_ext*(size_t)(6*6), cs_real_t, cs_alloc_mode);
+  CS_MALLOC_HD(xa, (size_t)n_edges*(size_t)(2*6*6), cs_real_t, cs_alloc_mode);
 
   /* Initialize arrays */
